+++ AGENTS.md
@@ -1,0 +1,1 @@
+See @CLAUDE.md for guidelines.
