+++ src/app/(i18n)/news/page.tsx
@@ -5,7 +5,7 @@ import { useLocale } from "next-intl";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
 import Template from "src/shared/ui/template";
 import NewsListSection from "src/widgets/news/list";
-import { useNewsPageQuery } from "src/entities/news/query/news.query";
+import { useNewsPageQuery } from "src/features/news/query/news.query";
 import { BigtabletSearchParams } from "src/shared/hooks/searchparams";
 import { Pagination } from "@bigtablet/design-system";
 import styles from "./style.module.scss";
