+++ src/app/(i18n)/recruit/[idx]/apply/page.tsx
@@ -12,9 +12,9 @@ import {
     applySchema,
     type ApplyFormValues,
 } from "src/features/recruit/apply/form/model/schema/apply.schema";
-import useEmailVerification from "src/features/recruit/apply/form/model/email/schema/verification";
+import useEmailVerification from "src/features/recruit/apply/form/model/email/use-email-verification";
 import ApplyForm from "src/features/recruit/apply/form/ui";
-import { useApplySubmit } from "src/features/recruit/apply/form/model/hook/apply.hook";
+import { useApplySubmit } from "src/features/recruit/apply/form/model/hook/use-apply-submit";
 
 const ApplyPage = () => {
     const { idx } = useParams<{ locale: string; idx: string }>();
