--- src/entities/blog/mutation/key.ts
@@ -1,3 +1,0 @@
-export const blogMutationKey = {
-    views: ["blog", "views"] as const,
-}
