--- src/entities/news/query/keys.ts
@@ -1,11 +1,0 @@
-export const newsQueryKeys = {
-    all: () => ["news"] as const,
-
-    lists: () => [...newsQueryKeys.all(), "list"] as const,
-
-    page: (page: number, size: number) =>
-        [...newsQueryKeys.all(), "page", page, size] as const,
-
-    detail: (idx: number) =>
-        [...newsQueryKeys.all(), "detail", idx] as const,
-};
