--- src/entities/talent/mutation/keys.ts
@@ -1,5 +1,0 @@
-export const talentMutationKeys = {
-    all: ["talent"] as const,
-    list: () => [...talentMutationKeys.all, "list"] as const,
-    create: () => [...talentMutationKeys.all, "create"] as const,
-};
