+++ src/features/blog/mutation/keys.ts
@@ -1,0 +1,3 @@
+export const blogMutationKeys = {
+	views: ["blog", "views"] as const,
+};
