+++ src/features/recruit/mutation/keys.ts
@@ -1,0 +1,4 @@
+export const recruitMutationKeys = {
+	all: ["recruit"] as const,
+	apply: () => [...recruitMutationKeys.all, "apply"] as const,
+};
