+++ src/features/talent/mutation/keys.ts
@@ -1,0 +1,5 @@
+export const talentMutationKeys = {
+	all: ["talent"] as const,
+	list: () => [...talentMutationKeys.all, "list"] as const,
+	create: () => [...talentMutationKeys.all, "create"] as const,
+};
