+++ src/features/upload/mutation/keys.ts
@@ -1,0 +1,4 @@
+export const uploadMutationKeys = {
+	all: ["upload"] as const,
+	gcp: () => [...uploadMutationKeys.all, "gcp"] as const,
+};
