--- src/shared/gcp/mutation/keys.ts
@@ -1,5 +1,0 @@
-export const Keys = {
-    gcp: {
-        upload: "/gcp",
-    },
-} as const;
